requirements.txt
python-docx>=0.8.11
pdf2image>=1.16.0
pdf2docx>=0.5.1
docx2pdf>=0.1.8
Pillow>=9.0.0
tkinterdnd2>=0.3.0
PyMuPDF>=1.23.0
